move code to common
