+++ client/src/DbtLanguageClient.ts
@@ -1,3 +1,4 @@
+import { TelemetryEvent } from 'dbt-language-server-common';
 import { Diagnostic, Disposable, OutputChannel, Uri, workspace, WorkspaceFolder } from 'vscode';
 import { LanguageClient, LanguageClientOptions, State, TransportKind, WorkDoneProgress } from 'vscode-languageclient/node';
 import { SUPPORTED_LANG_IDS } from './ExtensionClient';
@@ -5,11 +6,6 @@ import { ProgressHandler } from './ProgressHandler';
 import { PythonExtension } from './python/PythonExtension';
 import SqlPreviewContentProvider from './SqlPreviewContentProvider';
 import { TelemetryClient } from './TelemetryClient';
-
-interface TelemetryEvent {
-  name: string;
-  properties?: { [key: string]: string };
-}
 
 export class DbtLanguageClient implements Disposable {
   client: LanguageClient;
