+++ common/src/Interfaces.ts
@@ -1,0 +1,9 @@
+export interface TelemetryEvent {
+  name: string;
+  properties?: { [key: string]: string };
+}
+
+export interface PythonInfo {
+  path: string;
+  version?: string[];
+}
