--- common/src/Python.ts
@@ -1,4 +1,0 @@
-export interface PythonInfo {
-  path: string;
-  version?: string[];
-}
